(~) GdsFactoryExporter: Connection-Sektion in Partial extrahiert (500-Zeilen-Limit)
